tests/test_redis_functions.py
from redis import Redis
from testcontainers_demo.redis_functions import insert_random_key_value_pair, get_value
from random import Random
import pyperclip

async def test_get_and_set_value(test_redis: Redis):
    rng = Random(0)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


# this will generate a ton of unit tests
async def test_generate_tests():
    test_cases = []

    for i in range(0, 100):
        s = f'''
async def test_get_and_set_value_{i}(test_redis: Redis):
    rng = Random({i})
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1
'''
        test_cases.append(s)

    full_test_script = "\n".join(test_cases)
    pyperclip.copy(full_test_script)
    print("All test cases have been copied to the clipboard.")


# what follows is 100 unit tests that can be run in parallel

async def test_get_and_set_value_0(test_redis: Redis):
    rng = Random(0)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_1(test_redis: Redis):
    rng = Random(1)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_2(test_redis: Redis):
    rng = Random(2)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_3(test_redis: Redis):
    rng = Random(3)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_4(test_redis: Redis):
    rng = Random(4)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_5(test_redis: Redis):
    rng = Random(5)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_6(test_redis: Redis):
    rng = Random(6)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_7(test_redis: Redis):
    rng = Random(7)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_8(test_redis: Redis):
    rng = Random(8)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_9(test_redis: Redis):
    rng = Random(9)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_10(test_redis: Redis):
    rng = Random(10)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_11(test_redis: Redis):
    rng = Random(11)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_12(test_redis: Redis):
    rng = Random(12)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_13(test_redis: Redis):
    rng = Random(13)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_14(test_redis: Redis):
    rng = Random(14)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_15(test_redis: Redis):
    rng = Random(15)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_16(test_redis: Redis):
    rng = Random(16)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_17(test_redis: Redis):
    rng = Random(17)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_18(test_redis: Redis):
    rng = Random(18)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_19(test_redis: Redis):
    rng = Random(19)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_20(test_redis: Redis):
    rng = Random(20)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_21(test_redis: Redis):
    rng = Random(21)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_22(test_redis: Redis):
    rng = Random(22)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_23(test_redis: Redis):
    rng = Random(23)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_24(test_redis: Redis):
    rng = Random(24)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_25(test_redis: Redis):
    rng = Random(25)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_26(test_redis: Redis):
    rng = Random(26)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_27(test_redis: Redis):
    rng = Random(27)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_28(test_redis: Redis):
    rng = Random(28)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_29(test_redis: Redis):
    rng = Random(29)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_30(test_redis: Redis):
    rng = Random(30)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_31(test_redis: Redis):
    rng = Random(31)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_32(test_redis: Redis):
    rng = Random(32)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_33(test_redis: Redis):
    rng = Random(33)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_34(test_redis: Redis):
    rng = Random(34)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_35(test_redis: Redis):
    rng = Random(35)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_36(test_redis: Redis):
    rng = Random(36)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_37(test_redis: Redis):
    rng = Random(37)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_38(test_redis: Redis):
    rng = Random(38)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_39(test_redis: Redis):
    rng = Random(39)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_40(test_redis: Redis):
    rng = Random(40)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_41(test_redis: Redis):
    rng = Random(41)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_42(test_redis: Redis):
    rng = Random(42)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_43(test_redis: Redis):
    rng = Random(43)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_44(test_redis: Redis):
    rng = Random(44)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_45(test_redis: Redis):
    rng = Random(45)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_46(test_redis: Redis):
    rng = Random(46)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_47(test_redis: Redis):
    rng = Random(47)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_48(test_redis: Redis):
    rng = Random(48)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_49(test_redis: Redis):
    rng = Random(49)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_50(test_redis: Redis):
    rng = Random(50)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_51(test_redis: Redis):
    rng = Random(51)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_52(test_redis: Redis):
    rng = Random(52)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_53(test_redis: Redis):
    rng = Random(53)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_54(test_redis: Redis):
    rng = Random(54)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_55(test_redis: Redis):
    rng = Random(55)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_56(test_redis: Redis):
    rng = Random(56)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_57(test_redis: Redis):
    rng = Random(57)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_58(test_redis: Redis):
    rng = Random(58)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_59(test_redis: Redis):
    rng = Random(59)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_60(test_redis: Redis):
    rng = Random(60)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_61(test_redis: Redis):
    rng = Random(61)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_62(test_redis: Redis):
    rng = Random(62)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_63(test_redis: Redis):
    rng = Random(63)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_64(test_redis: Redis):
    rng = Random(64)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_65(test_redis: Redis):
    rng = Random(65)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_66(test_redis: Redis):
    rng = Random(66)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_67(test_redis: Redis):
    rng = Random(67)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_68(test_redis: Redis):
    rng = Random(68)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_69(test_redis: Redis):
    rng = Random(69)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_70(test_redis: Redis):
    rng = Random(70)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_71(test_redis: Redis):
    rng = Random(71)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_72(test_redis: Redis):
    rng = Random(72)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_73(test_redis: Redis):
    rng = Random(73)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_74(test_redis: Redis):
    rng = Random(74)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_75(test_redis: Redis):
    rng = Random(75)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_76(test_redis: Redis):
    rng = Random(76)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_77(test_redis: Redis):
    rng = Random(77)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_78(test_redis: Redis):
    rng = Random(78)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_79(test_redis: Redis):
    rng = Random(79)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_80(test_redis: Redis):
    rng = Random(80)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_81(test_redis: Redis):
    rng = Random(81)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_82(test_redis: Redis):
    rng = Random(82)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_83(test_redis: Redis):
    rng = Random(83)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_84(test_redis: Redis):
    rng = Random(84)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_85(test_redis: Redis):
    rng = Random(85)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_86(test_redis: Redis):
    rng = Random(86)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_87(test_redis: Redis):
    rng = Random(87)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_88(test_redis: Redis):
    rng = Random(88)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_89(test_redis: Redis):
    rng = Random(89)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_90(test_redis: Redis):
    rng = Random(90)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_91(test_redis: Redis):
    rng = Random(91)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_92(test_redis: Redis):
    rng = Random(92)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_93(test_redis: Redis):
    rng = Random(93)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_94(test_redis: Redis):
    rng = Random(94)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_95(test_redis: Redis):
    rng = Random(95)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_96(test_redis: Redis):
    rng = Random(96)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_97(test_redis: Redis):
    rng = Random(97)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_98(test_redis: Redis):
    rng = Random(98)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1


async def test_get_and_set_value_99(test_redis: Redis):
    rng = Random(99)
    assert len(test_redis.keys()) == 0
    k,v = await insert_random_key_value_pair(rng, test_redis)
    retrieved = await get_value(k, test_redis)
    assert v == retrieved
    assert len(test_redis.keys()) == 1
